requirements.txt
diffusers
transformers
accelerate
torchvision
pillow
streamlit